extract duplicated usages of file "selenide.properties" to a single class SelenideProperties
