+++ src/main/java/com/codeborne/selenide/impl/SelenideProperties.java
@@ -1,0 +1,7 @@
+package com.codeborne.selenide.impl;
+
+public class SelenideProperties extends PropertiesReader {
+  public SelenideProperties() {
+    super("selenide.properties");
+  }
+}
